+++ src/main.h
@@ -60,7 +60,7 @@ const gchar *main_get_version_string(void);
 
 gchar *main_get_argv_filename(const gchar *filename);
 
-void main_quit(void);
+gboolean main_quit(void);
 
 gboolean main_handle_filename(const gchar *locale_filename);
